Added create user account API with dependency update
