+++ src/main/java/com/wishlist/cst438project2/Common/Utils.java
@@ -1,16 +1,17 @@
 package com.wishlist.cst438project2.Common;
 
+import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Component;
 
-//@Component
+@Component
 public class Utils {
 
-    /*public static String encodePassword(String password) {
+    public static String encodePassword(String password) {
         password = BCrypt.hashpw(password, BCrypt.gensalt());
         return password;
     }
 
     public static boolean checkPassword(String password, String dbPassword) {
         return BCrypt.checkpw(password, dbPassword);
-    }*/
+    }
 }
